FInished OM homework, made quick change to CS project, and started some more work on OM project.

+++ OptMethods/project/code/wt.m
@@ -1,6 +1,6 @@
 function w = wt(x, n)
 	w = 0;
 	for i = 1:n
-		w = w + bitget(x, i);	
-	end
+		w = w + bitget(uint8(x), i);	
+    end
 end
